% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{tabset}
\alias{tabset}
\title{Tabset}
\usage{
tabset(..., id = NULL, selected = NULL, type = c("tabs", "pills"),
  stacked = FALSE, justified = FALSE)
}
\arguments{
\item{...}{HTML elements to include.}

\item{id}{Id of element. Can be accessed via \code{input$id} in server.}

\item{selected}{Name of the tab which is selected at start.}

\item{type}{One of "tabs" or "pills".}

\item{stacked}{Stacked tabset?}

\item{justified}{If \code{TRUE} tabs will use the full width.}
}
\description{
Like \link[shiny:tabsetPanel]{shiny::tabsetPanel} but with additional options, e.g. \code{justified}.
}
\examples{
if (interactive()) {
  library(shiny)

  ui <- bootstrapPage(
   tabset(tabPanel("Tab1", "You selected Tab1"),
          tabPanel("Tab2", "You selected Tab2"),
          justified = TRUE),

   tabset(tabPanel("Tab1", "You selected Tab1"),
          tabPanel("Tab2", "You selected Tab2"),
          type = "pills")
  )

  server <- function(input, output, session) {}
  shinyApp(ui, server)
}
}
