% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInputNoLabel.R
\name{textInputNoLabel}
\alias{textInputNoLabel}
\title{Text input without label}
\usage{
textInputNoLabel(id, value = "", placeholder = "")
}
\arguments{
\item{id}{Id of element. Can be accessed via \code{input$id} in server.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\description{
Same as \code{\link[shiny:textInput]{shiny::textInput()}} but without label.
}
